+++ src/main/kotlin/com/atlassian/performance/tools/aws/DefaultAmiNameRegistry.kt
@@ -1,0 +1,5 @@
+package com.atlassian.performance.tools.aws
+
+internal object DefaultAmiNameRegistry {
+    const val hvmSsdUbuntuXenial16_04Amd64Server20180912 = "ubuntu/images/hvm-ssd/ubuntu-xenial-16.04-amd64-server-20180912"
+}