+++ src/test/kotlin/com/atlassian/performance/tools/aws/FakeEc2.kt
@@ -1,0 +1,5 @@
+package com.atlassian.performance.tools.aws
+
+import com.amazonaws.services.ec2.AbstractAmazonEC2
+
+class FakeEc2 : AbstractAmazonEC2()